requirements.txt
openai>=1.0.0
anthropic>=0.40.0
google-genai>=1.0.0
python-dotenv>=1.0.0
matplotlib>=3.8.0
seaborn>=0.13.0
numpy>=1.26.0
